Improve import performance
